divison de responsabilidades de bd para cada modulo, hasta el momento todas estamas relacionadas con multiples columnas dentro de users.db
